requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-environ==0.11.2
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS Headers
django-cors-headers==4.3.1

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Celery for Background Tasks
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.0

# Rate Limiting
django-ratelimit==4.1.0

# HTTP Requests
requests==2.31.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
flake8==7.0.0
black==23.12.1

# Utilities
python-decouple==3.8

# Additional required packages
hiredis==2.3.2
